+++ src/exceptions.rs
@@ -8,4 +8,5 @@ pub enum Exception {
     NegativeSqrt,
     InvalidDataType,
     NativeFault,
+    IncorrectRegType,
 }